// src/app/interfaces/user-create.ts
export interface UserCreate {
  id: string;
  familyName: string;
  givenName: string;
  nickName: string;
  familyNameYomigana: string;
  givenNameYomigana: string;
  bday: number;
  gender: string;
  job: string;
  holiday: string;
  birthplace: string;
  nearestStation: string;
  hobby: string;
}